# refactor: tail-call version for flat_map_list