@polvalente 2021-09-20
Closing this because of a dumb mistake regarding the order of concatenation (which is what caused the major speed-up in the first place).

Refactoring to reverse each list + reversing the result worses the memory usage to double the usage with only 20% CPU gain, so it's not really worth it either.